fix: M82c defect pool closure — 9 legacy bugs (0.70.2 #6)
- sip_media.rs: drop redundant `as u8` (clippy cast_same_type)
- rex-sip: allow FFI false-positive clippy lints (arc_not_send_sync /
  type_complexity) + AudioBridge Default impl → zero warnings
- FilesPage.vue: S3 merged to single panel (drop right pane + divider)
- WorkspacePage.vue: F1 shortcut panel no-op on mobile (<768px)
- Bugs table: mark #1/#2/#3/#5/#7 already resolved in prior milestone
  (verified against current code), #4/#6/#8/#9 fixed this commit

+++ crates/rex-common/src/sip_media.rs
@@ -24,8 +24,9 @@ pub fn decode_media_frame(bytes: &[u8]) -> Vec<i16> {
     let n = bytes.len() / 2;
     let mut out = Vec::with_capacity(n);
     for i in 0..n {
-        let lo = bytes[i * 2] as u8;
-        let hi = bytes[i * 2 + 1] as u8;
+        // `bytes[*]` 已是 u8，无需 `as u8`（clippy `cast_same_type`）。
+        let lo = bytes[i * 2];
+        let hi = bytes[i * 2 + 1];
         out.push(i16::from_le_bytes([lo, hi]));
     }
     out
